Refactor: 폴더 구조 변경
기존 template 폴더에서 각자 page에 대한 컴포넌트 및 유틸함수를 관리했지만 너무 복잡하다 느껴 단순화시킴
모든 component및 lib 등은 공통으로 관리, 유의미한 네이밍으로 구분
