+++ src/lib/getImgValue.ts
