+++ src/types/imageTypes.ts
